requirements.txt
Django==4.0.4
python-decouple==3.6
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
djangorestframework==3.13.1
django-ckeditor==6.4.1
django-tables2==2.4.1
django-filter==21.1
